Organize code regarding eye and mouth offsets
There is no need to store same offset values in different places. One was stored in GraphicVectorMeshSet and the other in MeshPoolFS. The one in the MeshPoolFS would be used if development mode was on, and allowed changing the values in-game.
This does not come any benefits. GraphicVectorMeshSet is removed because there is no functional difference between it and Verse.GraphicMeshSet after removing the offset variables.